Added sweet alert on deletion

+++ src/pages/MyServices.jsx
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../Provider/AuthProvider";
 import { Link } from "react-router";
 import axios from "axios";
+import Swal from "sweetalert2";
 
 const MyServices = () => {
   const [myServices, setMyServices] = useState([]);
@@ -16,18 +17,45 @@ const MyServices = () => {
       .catch(err => console.log(err));
   }, [user?.email]);
 
-const handleDelete = (id)=>{
-  axios.delete(`http://localhost:3000/delete/${id}`)
-  .then(res =>{
-    console.log(res.data);
-    const filterData = myServices.filter(service => service._id != id)
-    setMyServices(filterData)
+  const handleDelete = (id) => {
+
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+        axios.delete(`http://localhost:3000/delete/${id}`)
+      .then(res => {
+        console.log(res.data);
+        if (res.data.deletedCount==1) {
+          const filterData = myServices.filter(service => service._id != id)
+        setMyServices(filterData)
+
+         Swal.fire({
+          title: "Deleted!",
+          text: "Your Service has been deleted.",
+          icon: "success"
+        });
+        }
+        
+
+      }).catch(err => {
+        console.log(err);
+
+      })
+       
+      }
+    });
+
+
     
-  }).catch(err =>{
-    console.log(err);
-    
-  })
-}
+  }
 
 
   return (
@@ -50,7 +78,7 @@ const handleDelete = (id)=>{
           <tbody className="bg-white">
             {myServices.map(service => (
               <tr key={service._id} className="hover:bg-gray-100 transition">
-                
+
                 <td>
                   <div className="flex items-center gap-3">
                     <div className="avatar">
@@ -81,7 +109,7 @@ const handleDelete = (id)=>{
                   <Link to={`/update-services/${service?._id}`}><button className="btn btn-primary btn-xs rounded-md px-4">
                     Edit
                   </button></Link>
-                  <button onClick={() =>handleDelete(service?._id)} className="btn btn-error btn-xs rounded-md px-4">
+                  <button onClick={() => handleDelete(service?._id)} className="btn btn-error btn-xs rounded-md px-4">
                     Delete
                   </button>
                 </td>
